+++ packages/server/graphql/mutations/helpers/publishToEmbedder.ts
@@ -2,5 +2,13 @@ import type {MessageToEmbedder} from '../../../../embedder/embedder'
 import getRedis from '../../../utils/getRedis'
 
 export const publishToEmbedder = (message: MessageToEmbedder) => {
-  return getRedis().xadd('embedderStream', 'MAXLEN', '~', 1000, '*', 'msg', JSON.stringify(message))
+  return getRedis().xadd(
+    'embedMetadataStream',
+    'MAXLEN',
+    '~',
+    1000,
+    '*',
+    'msg',
+    JSON.stringify(message)
+  )
 }
